Allow to ask for memory with 0-side padding. (#3648)
In the render pipeline, we promise 16 floats before the pointer passed to
each stage to be readable.

Because of implementation details of the pre-memory-manager image
allocation, that was always the case before. Once we started using the
memory manager, this is no longer the case. This PR introduces the
possibility to explicitly ask for such padding and uses it in the render
pipeline.

(cherry picked from commit af7bb4c7c604c931a749fa60c35b7f0405d1caad)